refactor: merge components and apply vue query
